+++ packages/core/src/Filter.ts
@@ -1,5 +1,5 @@
 export interface FilterInterface {
-  attribute?(name: string, value: string): string;
+  attribute?(name: string, value: any): any;
   node?(name: string, node: HTMLElement): HTMLElement | null;
 }
 
@@ -7,7 +7,7 @@ export default class Filter implements FilterInterface {
   /**
    * Filter and clean an HTML attribute value.
    */
-  attribute(name: string, value: string): string {
+  attribute(name: string, value: any): any {
     return value;
   }